% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/otherFunctions.R
\name{annotatePreCalculatedGRs}
\alias{annotatePreCalculatedGRs}
\title{Create a GRList with a reference GR and a sample GR annotated with GRanges provided in arguments.}
\usage{
annotatePreCalculatedGRs(
  grSetRep,
  grRef,
  myRefLabel,
  stringSet,
  grLScores = NULL,
  grLDistance = NULL,
  useSummitPMFlanking = T,
  flankingSize = 150
)
}
\arguments{
\item{grSetRep}{a GenomicRange with the merged replicates to annotate}

\item{grRef}{a GenomicRange with the (merged) reference to annotate}

\item{myRefLabel}{the name to give to the ref}

\item{stringSet}{the name to give to the set}

\item{grLScores}{a GRangeList with annotations for which you would like to know the best score overlapping the GRanges of experiments}

\item{grLDistance}{a GRangeList with annotations for which you would like to know the distance to the closest item}

\item{useSummitPMFlanking}{logical value to specify if you prefer to use the full region of the peak or only the summit +/- the `flankingSize`}

\item{flankingSize}{integer value used to extend the summit if `useSummitPMFlanking` is TRUE}
}
\value{
A list with:
`stringSet` a string with all samples other than the reference separated by comma
`myGRs` a GRangeList of 2 elements which are
      1. The peaks which are shared by all samples of mySet
      2. The peaks which are shared by all samples of myRef
`nameOfRef` the name of the reference(s) used
`namesOfGrLScores` the names of grLScores
`namesOfGrLDistance` the names of grLDistance
`useSummitPMFlanking` the value of useSummitPMFlanking which was used
`flankingSize` the value of flankingSize which was used
}
\description{
Create a GRList with a reference GR and a sample GR annotated with GRanges provided in arguments.
}
\details{
The GRanges of myGRs will have in mcols a new information which is
bestNameOfTheGRScore for each GR in grLScores and
distanceToTheNearestNameOfTheGR for each GR in grLDistance
}
