% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plottingFunctions.R
\name{annotateWithCate}
\alias{annotateWithCate}
\title{Add annotations to the list to be able to plot categories}
\usage{
annotateWithCate(myGRAndAttributes, thresholdsForgrLScores,
  thresholdsForgrLDistance)
}
\arguments{
\item{myGRAndAttributes}{Should be the output of \link[analysePeaks]{createMyGRs}}

\item{thresholdsForgrLScores}{a list with the threshold to use to make categories for the features where score is important}

\item{thresholdsForgrLDistance}{a list with the threshold to use to make categories for the features where distance is important}
}
\value{
a list like myGRAndAttributes but with more attributes.
}
\description{
Add annotations to the list to be able to plot categories
}
