% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plottingFunctions.R
\name{barplotFromNamedVector}
\alias{barplotFromNamedVector}
\title{Plot a barplot from a named vector}
\usage{
barplotFromNamedVector(
  data,
  colorsForGraphs = NULL,
  orderForCategory1 = NULL,
  legend = T,
  putLegendToTheTopRightBorder = T,
  args.legend = list(x = "topright", inset = c(-0.2, 0), bg = "white"),
  las = 2,
  ...
)
}
\arguments{
\item{data}{a named vector where values are number and names are category1(which can have dot).category2(no dot possible here)}

\item{colorsForGraphs}{a named vector with the colors which should be used in the barplot (names should correspond to category2)}

\item{orderForCategory1}{a vector with the names of the category1 in the order it should be plotted}

\item{legend}{a logical to precise if the legend should be added (default is TRUE)}

\item{putLegendToTheTopRightBorder}{a logical to precise if the legend should be until the extremity of the page (default is TRUE)}

\item{args.legend}{list of arguments for the legend (default is topright inset=(-0.2, 0), bg= "white")}

\item{las}{a numeric value between 0 and 3 to specify the style of axis label (default is 2)}

\item{...}{other arguments for barplot}
}
\value{
invisible x values for the barplot
}
\description{
Plot a barplot from a named vector
}
\details{
Will plot a barplot with one bar per category1
}
\examples{
topTs <- c(5, 10, 15, 20) * 1e3
colorsForGraphs<-grDevices::rainbow(2+length(topTs))
names(colorsForGraphs)<-c(paste0("overlapTop",sort(topTs/1e3),"k"),"overlap","specific")
v <- c(15809, rep(5000, 4), c(9316, 4762, 4327, 3831, 5051, 36638))
names(v) <- paste(c(rep("ChIP1", 5), rep("ChIP2", 6)),
                 c(rep(paste0("overlap", c("", "Top5k", "Top10k", "Top15k", "Top20k")), 2),
                   "specific"),
                 sep = ".")
barplotFromNamedVector(v, colorsForGraphs = colorsForGraphs)
}
