% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/otherFunctions.R
\name{cateNamesFromThreshold}
\alias{cateNamesFromThreshold}
\title{Give category names from thresholds}
\usage{
cateNamesFromThreshold(thresholdValues, lastCateName, unit)
}
\arguments{
\item{thresholdValues}{a numeric vector sorted in decreasing order which define the borders of categories}

\item{lastCateName}{a string which give the name of the category when it is below the last value of `thresholdValues`}

\item{unit}{a string which will be pasted after the threshold values}
}
\description{
Give category names from thresholds
}
