% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plottingFunctions.R
\name{createMyGRs}
\alias{createMyGRs}
\title{Create a GRList with a reference GR and a sample GR annotated with GRanges provided in arguments}
\usage{
createMyGRs(e, ovF, allExpe, grLScores = NULL, grLDistance = NULL,
  nameOfRef = "ChIP", useSummitPMFlanking = T, flankingSize = 150)
}
\arguments{
\item{e}{name of the experiment usually sample1____sample2(____sample3 etc...)}

\item{ovF}{a list which contains the filtered overlaps between the samples. Should at least contains a item named e.}

\item{allExpe}{a GRangeList with at least each sample in `e`}

\item{grLScores}{a GRangeList with annotations for which you would like to know the best score overlapping the GRanges of experiments}

\item{grLDistance}{a GRangeList with annotations for which you would like to know the distance to the closest item}

\item{nameOfRef}{among the samples which one is the reference, the others will be merged as "replicates"}

\item{useSummitPMFlanking}{logical value to specify if you prefer to use the full region of the peak or only the summit +/- the `flankingSize`}

\item{flankingSize}{integer value used to extend the summit if `useSummitPMFlanking` is TRUE}
}
\value{
A list with:
`stringSet` a string with all samples other than the reference separated by comma
`myGRs` a GRangeList of 2 elements which are
      1. The peaks which are shared by all samples which are considered as "replicates"
      2. The peaks of the reference
`nameOfRef` the name of the reference used
`namesOfGrLScores` the names of grLScores
`namesOfGrLDistance` the names of grLDistance
`useSummitPMFlanking` the value of useSummitPMFlanking which was used
`flankingSize` the value of flankingSize which was used
}
\description{
Create a GRList with a reference GR and a sample GR annotated with GRanges provided in arguments
}
\details{
The GRanges of myGRs will have in mcols a new information which is
bestNameOfTheGRScore for each GR in grLScores and
distanceToTheNearestNameOfTheGR for each GR in grLDistance
The first item of myGRs will also have inUniqueRef which is the index of the merged item of the Reference.
The second item of myGRs will also have inNoneOfTheSet
}
