% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plottingFunctions.R
\name{plotCateComparisonSetAndChIP}
\alias{plotCateComparisonSetAndChIP}
\title{Plot barplots for the scores or the distance of features in the Set and in the Reference
As well as when it is in Set and in the Set but not in the Reference and vice-versa}
\usage{
plotCateComparisonSetAndChIP(myGRAndAttributes, thresholdsForgrLScores,
  thresholdsForgrLDistance)
}
\arguments{
\item{myGRAndAttributes}{Should be the output of \link[analysePeaks]{createMyGRs}}

\item{thresholdsForgrLScores}{a list with the threshold to use to make categories for the features where score is important}

\item{thresholdsForgrLDistance}{a list with the threshold to use to make categories for the features where distance is important}
}
\value{
Plot barplots but do not return anything
}
\description{
Plot barplots for the scores or the distance of features in the Set and in the Reference
As well as when it is in Set and in the Set but not in the Reference and vice-versa
}
