% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plottingFunctions.R
\name{plotClassicalHistogramForMyGRs}
\alias{plotClassicalHistogramForMyGRs}
\title{Plot histograms for the scores or the distance of features in the Set and in the Reference
As well as when it is in Set and in the Set but not in the Reference and vice-versa}
\usage{
plotClassicalHistogramForMyGRs(myGRAndAttributes)
}
\arguments{
\item{myGRAndAttributes}{Should be the output of \link[analysePeaks]{createMyGRs} or \link[analysePeaks]{createMyGRsUsingSimpleOverlap}}
}
\value{
Plot histograms but do not return anything
}
\description{
Plot histograms for the scores or the distance of features in the Set and in the Reference
As well as when it is in Set and in the Set but not in the Reference and vice-versa
}
