% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plottingFunctions.R
\name{plotPairwiseComparison}
\alias{plotPairwiseComparison}
\title{Plot different plots to illustrate a comparison between 2 overlaps}
\usage{
plotPairwiseComparison(name1, name2, ovF, allExpe, step = 5000, fontsize = 10)
}
\arguments{
\item{name1}{the name of the first experiment to use}

\item{name2}{the name of the second experiment to use}

\item{ovF}{the list with at least the filtered overlap or the dataframe with cluster info for name1 and name2}

\item{allExpe}{a GRanges list with at least name1 and name2 items with a dataframe with a column called score}

\item{step}{a integer which will be used to bin the items of name1 or name2 into categories (default is 5000).}

\item{fontsize}{base fontsize for the heatmaps (default is 10)}
}
\value{
Will do a lot of plots but do not return anything.
}
\description{
Plot different plots to illustrate a comparison between 2 overlaps
}
\details{
`ovF` can be obtained by putting into a list the result of \link[usefulLDfunctionsGR]{filterByScore},
The name of the item in the list should be name1____name2
}
