% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/otherFunctions.R
\name{writeTablesFromMyGRsAndAnnot}
\alias{writeTablesFromMyGRsAndAnnot}
\title{Export myGRs with all annotations in tab-delimited text files.}
\usage{
writeTablesFromMyGRsAndAnnot(
  myGRAndAttributes,
  outputDirectory,
  useOriginalNarrowPeak = F
)
}
\arguments{
\item{myGRAndAttributes}{Should be the output of \link[analysePeaks]{createMyGRs} or \link[analysePeaks]{createMyGRsUsingSimpleOverlap}}

\item{outputDirectory}{a path to a folder where will be written the text files.}

\item{useOriginalNarrowPeak}{logical to specify if instead of the coordinates in myGRAndAttributes which are by default the summit +/-150bp. You want the original coordinates as in the input (default is F).}
}
\value{
Does not return anything but write 2 files into the `outputDirectory`:
Ref_table.txt
Set_table.txt
}
\description{
Export myGRs with all annotations in tab-delimited text files.
}
\details{
The coordinates are 1-based closed.
}
