% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plottingFunctions.R
\name{plotAllPheatmapsFor2CategoriesFromMyGR}
\alias{plotAllPheatmapsFor2CategoriesFromMyGR}
\title{Plot pheatmaps for categories of 2 specific features for a Set and a Ref}
\usage{
plotAllPheatmapsFor2CategoriesFromMyGR(myGRs, nameOfColWithCate1,
  cateNames1, what1, nameOfColWithCate2, cateNames2, what2, stringSet,
  nameOfRef)
}
\arguments{
\item{myGRs}{a GRangeList of 2 elements which are the set and the ref and should have 2 meta columns `nameOfColWithCate1` and `nameOfColWithCate2`}

\item{nameOfColWithCate1}{a string which gives the column to display}

\item{cateNames1}{the names of the categories, in the good order, for the first feature}

\item{what1}{the name of the first measure}

\item{nameOfColWithCate2}{a string which gives the column to display for the first feature}

\item{cateNames2}{the names of the categories, in the good order, for the second feature}

\item{what2}{the name of the second measure}

\item{stringSet}{the name of the set for the second feature}

\item{nameOfRef}{the name of the reference}
}
\value{
Plot but do not return anything
}
\description{
Plot pheatmaps for categories of 2 specific features for a Set and a Ref
}
