% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plottingFunctions.R
\name{plotCateComparisonSetAndRef}
\alias{plotCateComparisonSetAndRef}
\title{Plot barplots and pheatmaps for the scores or the distance of features in the Set and in the Reference
As well as when it is in Set and in the Set but not in the Reference and vice-versa}
\usage{
plotCateComparisonSetAndRef(myGRAndAttributes)
}
\arguments{
\item{myGRAndAttributes}{Should be the output of \link[analysePeaks]{annotateWithCate}}
}
\value{
Plot barplots and pheatmaps but do not return anything
}
\description{
Plot barplots and pheatmaps for the scores or the distance of features in the Set and in the Reference
As well as when it is in Set and in the Set but not in the Reference and vice-versa
}
