% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plottingFunctions.R
\name{plotCateComparisonSetAndRef}
\alias{plotCateComparisonSetAndRef}
\title{Plot barplots and pheatmaps for the scores or the distance of features in the Set and in the Reference
As well as when it is in Set and in the Set but not in the Reference and vice-versa}
\usage{
plotCateComparisonSetAndRef(
  myGRAndAttributes,
  fontsize = 10,
  plotBarPlots = TRUE,
  only5categoriesBP = FALSE,
  plotPheatmaps = TRUE,
  display_numbers = TRUE,
  whichPheatmaps = 1:5,
  allCates = NULL,
  plotProportion = FALSE
)
}
\arguments{
\item{myGRAndAttributes}{Should be the output of \link[analysePeaks]{annotateWithCate}}

\item{fontsize}{base fontsize for the heatmaps (default is 10)}

\item{plotBarPlots}{logical whether to plot the barplots (default is TRUE)}

\item{only5categoriesBP}{logical whether to plot only the barplots with 5 categories (default is FALSE).}

\item{plotPheatmaps}{logical whether to plot the pheatmaps (default is TRUE)}

\item{display_numbers}{logical whether to display the numbers in the pheatmaps (default is TRUE)}

\item{whichPheatmaps}{a vector containing the pheatmaps to plot between 1 to 5 with:
1=shared by all Set and at least one Ref or shared by all Set and Ref
2=shared by all Set none of the Ref or shared by all Set no Ref
3=shared by all Set: all or all shared by all Set
4=shared by all Ref in not a single of the set or Ref in not a single of the set
5=shared by all Ref: all or all Ref
(default is 1:5)}

\item{allCates}{a vector of string with the categories to plot, if NULL all categories in `myGRAndAttributes` are used (default is NULL)}

\item{plotProportion}{logical whether to display the pheatmap 1, 2, and 4 should be plot as proportion of pheatmap 3 and 5 (default is FALSE)}
}
\value{
Plot barplots and pheatmaps but do not return anything
}
\description{
Plot barplots and pheatmaps for the scores or the distance of features in the Set and in the Reference
As well as when it is in Set and in the Set but not in the Reference and vice-versa
}
